+++ js/bulk-move.js
@@ -10,7 +10,7 @@
 /*jslint browser: true, devel: true*/
 /*global BULK_MOVE, jQuery, document, postboxes, pagenow*/
 jQuery(document).ready(function () {
-    jQuery('button[value="bulk-move-tags"], button[value="bulk-move-cats"]').click(function () {
+    jQuery('button[value="bulk-move-tags"], button[value="bulk-move-cats"], button[value="bulk-move-category-by-tag"]').click(function () {
         return confirm(BULK_MOVE.msg.move_warning);
     });